% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chiSquareTest.R
\name{chiSquareTest}
\alias{chiSquareTest}
\title{Contingency matrix form datasets.}
\usage{
chiSquareTest(productionSteps, waferClustering, operation, cluster, equipment,
  threshold = mean(waferClustering$faults))
}
\arguments{
\item{productionSteps:}{a data.frame where the first column is assumed to contain the lot, the second one the equipment and the
third the operations during the production chain}

\item{waferClustering:}{a data.frame where the first column is assumed to contain the lot, the second one the wafer ID and the
third the number of faults and the last the cluster as returned by SOM algorithm.}

\item{operation:}{the operation (ie. the production step to be tested)}

\item{cluster:}{the cluster to investigate}

\item{equipment:}{the equipment whose yield is to be tested}

\item{threshold:}{the maximum nuber of fault a good wafer could have. Default is the average of the waferClustering data.frame.}
}
\value{
either an object as returned by chisq.test or a string explaining why the test could not have been performed
}
\description{
This function creates a contingency matrix from two dataset, containing the production steps and the result of wafer clustering,
respectively.
The test deals with the probability of a particular equipment to show a lower yield than the average one. The higher is
the value of the test, the most likely is the equipment to be faulty.
}

