% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createContingencyMatrix.R
\name{createContingencyMatrix}
\alias{createContingencyMatrix}
\title{Contingency matrix form datasets.}
\usage{
createContingencyMatrix(productionSteps, waferClustering, operation, cluster,
  equipment, threshold = mean(waferClustering$faults))
}
\arguments{
\item{productionSteps:}{a data.frame where the first column is assumed to contain the lot, the second one the equipment and the
third the operations during the production chain}

\item{waferClustering:}{a data.frame where the first column is assumed to contain the lot, the second one the wafer ID and the
third the number of faults and the last the cluster as returned by SOM algorithm.}

\item{operation:}{the operation (ie. the production step to be tested)}

\item{cluster:}{the cluster to investigate}

\item{equipment:}{the equipment whose yield is to be tested}

\item{threshold:}{the maximum nuber of fault a good wafer could have. Default is the average of the waferClustering data.frame.}
}
\value{
the 2x2 contingency matrix
}
\description{
This function creates a contingency matrix from two dataset, containing the production steps and the result of wafer clustering,
respectively.
The contingency matrix is a 2x2 numeric matrix M.
M11 = bad wafer processed by the selected equipment
M12 = bad wafer processed by other equipments
M21 = good wafer processed by the selected equipment
M22 = bad wafer processed by other equipments
}

